+++ horovod/run/__init__.py
@@ -13,4 +13,4 @@
 # limitations under the License.
 # =============================================================================
 
-from .run import run
+from .run import run, run_controller
